refactor: give numerator relations its own dir
