+++ src/components/config-tabs/ConfigTabs.js
@@ -2,12 +2,12 @@ import { CircularLoader } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import styles from './ConfigTabs.module.css'
+import { NumeratorRelations } from './numerator-relations/index.js'
 import { DenominatorRelations } from './tab-contents/DenominatorRelations.js'
 import { Denominators } from './tab-contents/Denominators.js'
 import { ExternalDataComparison } from './tab-contents/ExternalDataComparison.js'
 import { NumeratorGroups } from './tab-contents/NumeratorGroups.js'
 import { NumeratorParameters } from './tab-contents/NumeratorParameters.js'
-import { NumeratorRelations } from './tab-contents/NumeratorRelations.js'
 import { Numerators } from './tab-contents/Numerators.js'
 
 function Tabs({ loading, configurations, mappedNumerators }) {
