+++ src/components/config-tabs/numerator-relations/index.js
@@ -1,0 +1,1 @@
+export { NumeratorRelations } from './NumeratorRelations.js'
